% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_pkg.R
\name{show_pkg}
\alias{show_pkg}
\title{Show Package Details}
\usage{
show_pkg(package = NULL, lib_path = NULL, return_value = c("print", "data"))
}
\arguments{
\item{package}{Name of package.}

\item{lib_path}{Default is \code{.libPaths()}, custom path to packages 
can be specified.}

\item{return_value}{One of "print" (prints out the results to the 
R console) or "data" (returns the results as a dataframe).}
}
\description{
Prints or returns information on a certain package.
}
\examples{
\dontrun{
show_pkg("MASS")
}
}
